requirements.txt
PyQt
sys
PyQt-tools
